+++ src/address.rs
@@ -1,19 +1,34 @@
 use clap::Subcommand;
+use serde::Deserialize;
 
-#[derive(Subcommand)]
+use crate::{get, get_auth};
+
+#[derive(Debug, Subcommand)]
 pub enum Address {
     /// Get information about the availability of an address
-    IsAvailable,
+    IsAvailable {
+        /// Address to get availability of
+        address: String,
+    },
     /// Get the expiration date for an address
-    GetExpiry,
+    GetExpiry {
+        /// Address to get availability of
+        address: String,
+    },
     /// Get limited (public) information about an address (no auth required)
-    GetPublicInfo,
-    ///Get comprehensive information about an address
-    GetInfo,
+    GetPublicInfo {
+        /// Address to get availability of
+        address: String,
+    },
+    /// Get comprehensive information about an address
+    GetInfo {
+        /// Address to get availability of
+        address: String,
+    },
 }
 
 impl Address {
-    pub fn process(&self, _address: &Option<String>) {
+    pub fn process(&self, api_key: &str) -> Result<AddressResponse, reqwest::Error> {
         match self {
             Address::IsAvailable => todo!(),
             Address::GetExpiry => todo!(),